+++ packages/effector/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @virentia/effector
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @virentia/core@0.7.1
+
 ## 0.4.0
 
 ### Minor Changes