+++ packages/mutable/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @virentia/mutable
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @virentia/core@0.7.1
+
 ## 0.2.0
 
 ### Minor Changes