+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @virentia/react
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @virentia/core@0.7.1
+
 ## 0.3.0
 
 ### Minor Changes